Create "SOS"_Game
This program uses Java to create a "tic-tac-toe" style game but instead uses "S" and "O". The player must create an "SOS" pattern and the user with more instances of creating "SOS" wins. The main point of this program was to create a GUI while also implementing the code for the game.